+++ ui/components/modals/base/example.jsx
@@ -58,7 +58,7 @@ export let Modal = (props) => (
       {
         props.closeButton !== false ? (
           <ButtonIcon
-            className="slds-modal__close slds-button_icon-inverse"
+            className="slds-modal__close"
             iconClassName="slds-button__icon_large"
             symbol="close"
             assistiveText="Cancel and close"
@@ -83,13 +83,13 @@ Modal.defaultProps = {
 };
 
 export let ModalHeader = (props) => (
-  <div className={classNames('slds-modal__header', props.className)}>
+  <div className={classNames('slds-modal__header slds-text-align_left', props.className)}>
     {props.children}
   </div>
 );
 
 export let ModalHeaderDeprecated = (props) => (
-  <div className={classNames('slds-modal__header', props.className)}>
+  <div className={classNames('slds-modal__header slds-text-align_left', props.className)}>
     {props.closeButton !== false ? (
       <ButtonIcon
         className="slds-modal__close slds-button_icon-inverse"
@@ -136,7 +136,7 @@ export let Taglines = (props) => (
   <Backdrop>
     <Modal aria-labelledby="modal-heading-01">
       <ModalHeader>
-        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate">
+        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
           Modal header
         </h1>
         <p className="slds-m-top_x-small">
@@ -188,7 +188,7 @@ export let ModalSizes = (props) => (
       aria-labelledby="modal-heading-01"
     >
       <ModalHeader>
-        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate">
+        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
           Modal header
         </h1>
       </ModalHeader>
@@ -221,7 +221,7 @@ export let FullTaglines = (props) => (
       )}
       isFullSize={props.size === 'full'}>
       <ModalHeader>
-        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate">
+        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
           Modal header
         </h1>
         <p className="slds-m-top_x-small">
@@ -319,7 +319,7 @@ export let Directional = (props) => (
   <Backdrop>
     <Modal aria-labelledby="modal-heading-01">
       <ModalHeader>
-        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate">
+        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
           Modal header
         </h1>
       </ModalHeader>
@@ -416,7 +416,7 @@ export let HiddenFooter = (props) => (
   <Backdrop>
     <Modal aria-labelledby="modal-heading-01">
       <ModalHeader>
-        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate">
+        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
           Modal header
         </h1>
       </ModalHeader>
@@ -538,7 +538,7 @@ export let DisabledClose = (props) => (
   <Backdrop>
     <Modal aria-labelledby="modal-heading-01" closeButtonDisabled>
       <ModalHeader>
-        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate">
+        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
           Modal header
         </h1>
       </ModalHeader>
@@ -568,7 +568,7 @@ export let DeprecatedClose = (props) => (
   <Backdrop>
     <Modal aria-labelledby="modal-heading-01" closeButton={false}>
       <ModalHeaderDeprecated>
-        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate">
+        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
           Modal header
         </h1>
       </ModalHeaderDeprecated>
@@ -602,8 +602,8 @@ const defaultComponent = (
   <Backdrop>
     <Modal aria-labelledby="modal-heading-01">
       <ModalHeader>
-        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate">
-          Modal header
+        <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
+          This is long Modal header with a long and long and very long text which may slipped down to a new line
         </h1>
       </ModalHeader>
       <ModalContent className="slds-p-around_medium">